Avoid unhandled exception in dimension handling
In particular, making a request with TIME=CURRENT would prompt GeoServer
to try to evaluate the maximum time of an empty layer.  The fix is
simple - just check whether the MaxVisitor has a result before asking it
for one.  A similar fix is applied for finding the minimum elevation of
an elevation-enabled layer.